+++ paymentsheet/src/androidTest/java/com/stripe/android/paymentsheet/CustomerSheetTest.kt
@@ -8,6 +8,8 @@ import com.stripe.android.core.utils.urlEncode
 import com.stripe.android.customersheet.CustomerSheet
 import com.stripe.android.customersheet.CustomerSheetResult
 import com.stripe.android.customersheet.ExperimentalCustomerSheetApi
+import com.stripe.android.customersheet.PaymentOptionSelection
+import com.stripe.android.model.CardBrand
 import com.stripe.android.networktesting.NetworkRule
 import com.stripe.android.networktesting.RequestMatcher
 import com.stripe.android.networktesting.RequestMatchers
@@ -21,6 +23,7 @@ import com.stripe.android.paymentsheet.utils.CustomerSheetTestType
 import com.stripe.android.paymentsheet.utils.CustomerSheetTestTypeProvider
 import com.stripe.android.paymentsheet.utils.IntegrationType
 import com.stripe.android.paymentsheet.utils.IntegrationTypeProvider
+import com.stripe.android.paymentsheet.utils.PrefsTestStore
 import com.stripe.android.paymentsheet.utils.runCustomerSheetTest
 import org.junit.Rule
 import org.junit.Test
@@ -88,6 +91,55 @@ internal class CustomerSheetTest {
         networkRule.enqueueAttachRequests(customerSheetTestType)
 
         page.clickSaveButton()
+        page.clickConfirmButton()
+    }
+
+    @Test
+    fun testSavedCardReturnedInResultCallback() = activityScenarioRule.runCustomerSheetTest(
+        integrationType = integrationType,
+        customerSheetTestType = CustomerSheetTestType.AttachToSetupIntent,
+        resultCallback = { result ->
+            assertThat(result).isInstanceOf(CustomerSheetResult.Selected::class.java)
+
+            val selected = result as CustomerSheetResult.Selected
+
+            assertThat(selected.selection).isInstanceOf(PaymentOptionSelection.PaymentMethod::class.java)
+
+            val paymentMethodSelection = selected.selection as PaymentOptionSelection.PaymentMethod
+
+            val card = paymentMethodSelection.paymentMethod.card
+
+            assertThat(card?.last4).isEqualTo("4242")
+            assertThat(card?.brand).isEqualTo(CardBrand.Visa)
+        }
+    ) { context ->
+        context.scenario.onActivity {
+            PrefsTestStore(it).clear()
+        }
+
+        networkRule.enqueue(
+            retrieveElementsSessionRequest(),
+        ) { response ->
+            response.testBodyFromFile("elements-sessions-requires_payment_method.json")
+        }
+
+        networkRule.enqueue(
+            retrievePaymentMethodsRequest(),
+            cardPaymentMethodsParams(),
+        ) { response ->
+            response.testBodyFromFile("payment-methods-get-success.json")
+        }
+
+        networkRule.enqueue(
+            retrievePaymentMethodsRequest(),
+            usBankAccountPaymentMethodsParams(),
+        ) { response ->
+            response.testBodyFromFile("payment-methods-get-success-empty.json")
+        }
+
+        context.presentCustomerSheet()
+
+        page.clickSavedPaymentMethod(endsWith = "4242")
         page.clickConfirmButton()
     }