+++ paymentsheet/src/main/java/com/stripe/android/lpmfoundations/luxe/TransformSpecToElements.kt
@@ -77,7 +77,7 @@ internal class TransformSpecToElements(
                     arguments.shippingValues
                 )
                 is SepaMandateTextSpec -> it.transform(arguments.merchantName)
-                is PlaceholderSpec -> error("Placeholders should be processed before calling transform.")
+                is PlaceholderSpec -> null // Placeholders should be processed before calling transform.
                 is CashAppPayMandateTextSpec -> it.transform(arguments.merchantName)
                 is KlarnaMandateTextSpec -> it.transform(arguments.merchantName)
             }