+++ stripe-core/src/main/java/com/stripe/android/core/exception/StripeException.kt
@@ -63,8 +63,16 @@ abstract class StripeException(
                     message = throwable.message,
                     cause = throwable
                 )
-                else -> GenericStripeException(throwable)
+                else -> GenericStripeException(throwable, analyticsValue = analyticsValueForThrowable(throwable))
             }
+        }
+
+        private fun analyticsValueForThrowable(throwable: Throwable): String? {
+            val throwableClass = throwable.javaClass
+            if (throwableClass.name.startsWith("android.") || throwableClass.name.startsWith("java.")) {
+                return throwableClass.name
+            }
+            return null
         }
     }
 }